test: fold ignore config coverage
